MCHoverThrustEstimator: Always publish an estimate
Always publish an estimate even when not fusing measurements. This is
required as the land detector and the position controller need to
receive a hover thrust value.
Also remove the "armed" check as it can't be !_landed and !_armed at the
same time. And add a check to not run the estimator before we reach 1m
agl.